# back/openapi.json
{"openapi":"3.1.0","info":{"title":"Employee well-being monitoring API","version":"0.1.0"},"paths":{"/v1/register/managers":{"post":{"summary":"Register User","operationId":"register_user_v1_register_managers_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManagerRegisterBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/register/employees":{"post":{"summary":"Register User","operationId":"register_user_v1_register_employees_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeRegisterBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/login":{"post":{"summary":"Login User","operationId":"login_user_v1_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/instructions":{"get":{"summary":"Get Instruction","operationId":"get_instruction_v1_instructions_get","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstructionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/states":{"get":{"summary":"Get States","operationId":"get_states_v1_states_get","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/factors":{"get":{"summary":"Get Factors","operationId":"get_factors_v1_factors_get","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FactorsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/feedbacks/{feedback_date}":{"get":{"summary":"Get Instruction","operationId":"get_instruction_v1_feedbacks__feedback_date__get","parameters":[{"name":"feedback_date","in":"path","required":true,"schema":{"type":"string","format":"date","title":"Feedback Date"}},{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/feedbacks/{feedback_date}/morning":{"post":{"summary":"Save Morning","operationId":"save_morning_v1_feedbacks__feedback_date__morning_post","parameters":[{"name":"feedback_date","in":"path","required":true,"schema":{"type":"string","format":"date","title":"Feedback Date"}},{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MorningBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/feedbacks/{feedback_date}/evening":{"post":{"summary":"Save Evening","operationId":"save_evening_v1_feedbacks__feedback_date__evening_post","parameters":[{"name":"feedback_date","in":"path","required":true,"schema":{"type":"string","format":"date","title":"Feedback Date"}},{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EveningBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/participants":{"get":{"summary":"Get Participants","operationId":"get_participants_v1_participants_get","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParticipantResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/managers":{"get":{"summary":"Get Managers","operationId":"get_managers_v1_managers_get","parameters":[{"name":"authorization","in":"header","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManagerListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EmployeeRegisterBody":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"surname":{"type":"string","title":"Surname"},"birthdate":{"type":"string","format":"date","title":"Birthdate"},"phone":{"type":"string","title":"Phone"},"position":{"type":"string","title":"Position"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"manager_id":{"type":"integer","title":"Manager Id"}},"type":"object","required":["first_name","last_name","surname","birthdate","phone","position","email","password","manager_id"],"title":"EmployeeRegisterBody"},"EveningBody":{"properties":{"state_id":{"type":"integer","title":"State Id"},"factor_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Factor Id"}},"type":"object","required":["state_id","factor_id"],"title":"EveningBody"},"Factor":{"properties":{"id":{"type":"integer","title":"Id"},"value":{"type":"string","title":"Value"}},"type":"object","required":["id","value"],"title":"Factor"},"FactorCategory":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"factors":{"items":{"$ref":"#/components/schemas/Factor"},"type":"array","title":"Factors"}},"type":"object","required":["id","name","factors"],"title":"FactorCategory"},"FactorsResponse":{"properties":{"result":{"items":{"$ref":"#/components/schemas/FactorCategory"},"type":"array","title":"Result"}},"type":"object","required":["result"],"title":"FactorsResponse"},"FeedbackResponse":{"properties":{"morning":{"type":"boolean","title":"Morning"},"evening":{"type":"boolean","title":"Evening"}},"type":"object","required":["morning","evening"],"title":"FeedbackResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InstructionResponse":{"properties":{"result":{"type":"string","title":"Result"}},"type":"object","required":["result"],"title":"InstructionResponse"},"LoginBody":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginBody"},"LoginResponse":{"properties":{"result":{"type":"string","title":"Result"}},"type":"object","required":["result"],"title":"LoginResponse"},"ManagerListItem":{"properties":{"id":{"type":"integer","title":"Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"surname":{"type":"string","title":"Surname"},"company":{"type":"string","title":"Company"},"department":{"type":"string","title":"Department"}},"type":"object","required":["id","first_name","last_name","surname","company","department"],"title":"ManagerListItem"},"ManagerListResponse":{"properties":{"result":{"items":{"$ref":"#/components/schemas/ManagerListItem"},"type":"array","title":"Result"}},"type":"object","required":["result"],"title":"ManagerListResponse"},"ManagerRegisterBody":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"surname":{"type":"string","title":"Surname"},"birthdate":{"type":"string","format":"date","title":"Birthdate"},"phone":{"type":"string","title":"Phone"},"position":{"type":"string","title":"Position"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"company":{"type":"string","title":"Company"},"department":{"type":"string","title":"Department"}},"type":"object","required":["first_name","last_name","surname","birthdate","phone","position","email","password","company","department"],"title":"ManagerRegisterBody"},"MorningBody":{"properties":{"state_id":{"type":"integer","title":"State Id"}},"type":"object","required":["state_id"],"title":"MorningBody"},"ParticipantListItem":{"properties":{"id":{"type":"integer","title":"Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"surname":{"type":"string","title":"Surname"},"birthdate":{"type":"string","format":"date","title":"Birthdate"},"company":{"type":"string","title":"Company"},"position":{"type":"string","title":"Position"},"phone":{"type":"string","title":"Phone"},"email":{"type":"string","title":"Email"}},"type":"object","required":["id","first_name","last_name","surname","birthdate","company","position","phone","email"],"title":"ParticipantListItem"},"ParticipantResponse":{"properties":{"result":{"items":{"$ref":"#/components/schemas/ParticipantListItem"},"type":"array","title":"Result"}},"type":"object","required":["result"],"title":"ParticipantResponse"},"State":{"properties":{"id":{"type":"integer","title":"Id"},"value":{"type":"string","title":"Value"}},"type":"object","required":["id","value"],"title":"State"},"StatesResponse":{"properties":{"res":{"items":{"$ref":"#/components/schemas/State"},"type":"array","title":"Res"}},"type":"object","required":["res"],"title":"StatesResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}